Conflicts:
	bundles/org.eclipse.orion.client.ui/web/orion/editorView.js
	bundles/org.eclipse.orion.client.ui/web/orion/liveEditSession.js

+++ bundles/org.eclipse.orion.client.ui/web/orion/compare/resourceComparer.js
@@ -81,7 +81,7 @@ exports.DefaultDiffProvider = (function() {
 			}, function(){});
 		},
 		
-		resolveDiff: function(resource, compareTo, hasConflicts) {
+		resolveDiff: function(resource, compareTo, hasConflicts, ignoreWhitespace) {
 			this._hasConflicts = hasConflicts;
 			if(compareTo){
 				return this._resolveTwoFiles(compareTo, resource);
@@ -91,14 +91,15 @@ exports.DefaultDiffProvider = (function() {
 					return;
 				}
 				var that = this;
-				if(that._hasConflicts) {
-					return that._diffProvider.getDiffContent(resource).then(function(jsonData, secondArg) {
+				var ignoreWS = ignoreWhitespace ? "true" : "false";
+				return that._diffProvider.getDiffContent(resource, {ignoreWS: ignoreWS}).then(function(jsonData) {
+					if (that._hasConflicts) {
 						that._diffContent = jsonData.split("diff --git")[1]; //$NON-NLS-0$
-						return that._resolveComplexFileURL(resource);
-					}, function(){});
-				} else {// We do not need to get diff contents from server if there is no conflicts. We use client side diff here.
+					} else {
+						that._diffContent = jsonData;
+					}
 					return that._resolveComplexFileURL(resource);
-				}
+				}, function(){});
 			}
 		}
 	};
@@ -116,6 +117,8 @@ CompareStyler.prototype = {
 									 false /*bug 378193*/);
 	}
 };
+var SAVE_EMBEDDED = true;
+//SAVE_EMBEDDED = (new URL(window.location.href).query.get("save") === "true");
 
 exports.ResourceComparer = (function() {
 	function ResourceComparer (serviceRegistry, commandRegistry, options, viewOptions) {
@@ -126,6 +129,20 @@ exports.ResourceComparer = (function() {
 		this._searchService = this._registry.getService("orion.core.search"); //$NON-NLS-0$
 		this._progress = this._registry.getService("orion.page.progress"); //$NON-NLS-0$
 		this.setOptions(options, true);
+		this._inputManagers = [];
+		viewOptions.preCreate = this._initInputManagers.bind(this);
+		viewOptions.postCreate = function () {
+			this._inputManagers.forEach(function(inputManager) {
+				if(inputManager.manager){
+					var editor = this._compareView.getWidget().getEditors()[inputManager.manager._editorIndex];
+					editor.addEventListener("DirtyChanged", function(evt) { //$NON-NLS-0$
+						inputManager.manager.setDirty(editor.isDirty());
+					});
+				}
+			}.bind(this));
+		}.bind(this);
+		viewOptions.onInputChanged = this._inputChanged.bind(this);
+		viewOptions.onSave = this.save.bind(this);
 		if(options.toggleable) {
 			this._compareView = new mCompareView.toggleableCompareView(options.type === "inline" ? "inline" : "twoWay", viewOptions); //$NON-NLS-2$ //$NON-NLS-1$ //$NON-NLS-0$
 		} else if(options.type === "inline") { //$NON-NLS-0$
@@ -134,6 +151,9 @@ exports.ResourceComparer = (function() {
 			this._compareView = new mCompareView.TwoWayCompareView(viewOptions);
 		}
 		this._compareView.getWidget().setOptions({ignoreWhitespace: options.ignoreWhitespace});
+		this._compareView.getWidget().setOptions({diffProvider: options.diffProvider});
+		this._compareView.getWidget().setOptions({resource: options.resource});
+		this._compareView.getWidget().setOptions({titleIds: this.options.saveLeft ? this.options.saveLeft.titleIds : null });
 		this._compareView.getWidget().setOptions({extCmdHolder: this});
 		if(!viewOptions.highlighters){
 			this._compareView.getWidget().setOptions({highlighters: [new CompareStyler(serviceRegistry), new CompareStyler(serviceRegistry)]});
@@ -145,16 +165,7 @@ exports.ResourceComparer = (function() {
 			this._compareView.getWidget().setOptions({newFile: {readonly: this._checkReadonly(options.readonly, true)}});
 		}
 		this.initExtCmds();
-		this._initInputManagers();
 		this._compareView.getWidget().initEditors( messages['fetching...']);
-		this._inputManagers.forEach(function(inputManager) {
-			if(inputManager.manager){
-				var editor = this._compareView.getWidget().getEditors()[inputManager.manager._editorIndex];
-				editor.addEventListener("DirtyChanged", function(evt) { //$NON-NLS-0$
-					inputManager.manager.setDirty(editor.isDirty());
-				});
-			}
-		}.bind(this));
 	}
 	ResourceComparer.prototype = {
 		_clearOptions: function(){
@@ -162,8 +173,14 @@ exports.ResourceComparer = (function() {
 		},
 		_checkReadonly: function(readonlyFlag, isLeft){
 			if(isLeft){//Legacy: If not defined, left side readonly is false
+				if(this.options.saveLeft && !SAVE_EMBEDDED) {
+					return true;
+				}
 				return typeof readonlyFlag === "undefined" ? false : readonlyFlag; //$NON-NLS-0$
 			} else {//Legacy: If not defined, right side readonly is true
+				if(this.options.saveRight && !SAVE_EMBEDDED) {
+					return true;
+				}
 				return typeof readonlyFlag === "undefined" ? true : readonlyFlag; //$NON-NLS-0$
 			}
 		},
@@ -171,25 +188,32 @@ exports.ResourceComparer = (function() {
 			return (editorIndex === 1 ? this._compareView.getWidget().options.newFile : this._compareView.getWidget().options.oldFile);
 		},
 		_createInputManager: function() {
-			return new mInputManager.InputManager({
+			var im = new mInputManager.InputManager({
 						serviceRegistry: this._registry,
 						fileClient: this._fileClient,
 						progressService: this._progress
 			});
+			im.getReadOnly = function() {
+				return false;
+			};
+			return im;
+		},
+		_saveCmdVisible: function() {
+			return this._compareView.getWidget().type === "twoWay";
 		},
 		_initInputManagers: function() {
 			this._inputManagers = [{}, {}];
 			//We only create input managers when it is a non toggleable side by side compare widget
-			if(!this.options.toggleable && this._compareView.getWidget()._uiFactory && this._compareView.getWidget().type === "twoWay") { //$NON-NLS-0$
+			if(this._compareView.getWidget()._uiFactory && this._compareView.getWidget().type === "twoWay") { //$NON-NLS-0$
 				//Create the right hand side input manager
 				if(!this._checkReadonly(this.options.readonlyRight)) {
 					this._inputManagers[0].manager = this._createInputManager();
-					this._initInputManager(this._inputManagers[0].manager, 0, this._compareView.getWidget()._uiFactory.getActionDivId());
+					this._initInputManager(this._inputManagers[0].manager, 0, this.options.saveRight ? this.options.saveRight.saveCmdContainerId : this._compareView.getWidget()._uiFactory.getActionDivId(), !!this.options.saveRight);
 				}
 				//Create the left hand side input manager
 				if(!this._checkReadonly(this.options.readonly, true)) {
 					this._inputManagers[1].manager =  this._createInputManager();
-					this._initInputManager(this._inputManagers[1].manager, 1, this._compareView.getWidget()._uiFactory.getActionDivId(true));
+					this._initInputManager(this._inputManagers[1].manager, 1, this.options.saveLeft ? this.options.saveLeft.saveCmdContainerId : this._compareView.getWidget()._uiFactory.getActionDivId(true), !!this.options.saveLeft);
 				}
 			}
 			var that = this;
@@ -208,20 +232,22 @@ exports.ResourceComparer = (function() {
 							navToolbarId: "pageNavigationActions", //$NON-NLS-0$
 							textSearcher: localSearcher
 						});
-						var saveCmdId = inputManager.manager._editorIndex === 1 ? "orion.compare.save.left" : "orion.compare.save.right"; //$NON-NLS-1$ //$NON-NLS-0$
-						commandGenerator.generateSimpleEditorCommands(editor, saveCmdId);
+						var saveCmdId = inputManager.manager._editorIndex === 1 ? (that.options.saveLeft ? that.options.saveLeft.saveCmdId : "orion.compare.save.left") : 
+													(that.options.saveRight ? that.options.saveRight.saveCmdId : "orion.compare.save.right"); //$NON-NLS-1$ //$NON-NLS-0$
+						commandGenerator.generateSimpleEditorCommands(editor, saveCmdId, function() {return that._saveCmdVisible();}, 2000);
 						return keyBindings;
 					};
 					this._getFileOptions(inputManager.manager._editorIndex).keyBindingFactory = keyBindingFactory;
 				}
 			}.bind(this));
 		},
-		_initInputManager: function(inputManger, editorIndex, actionBarId){
+		_initInputManager: function(inputManger, editorIndex, actionBarId, embedded){
 			var that = this;
 			objects.mixin(inputManger, {
 				filePath: "",
 				_editorIndex: editorIndex,
 				_actionBarId: actionBarId,
+				embedded: embedded,
 				getInput: function() {
 					return this.filePath;
 				},
@@ -229,12 +255,35 @@ exports.ResourceComparer = (function() {
 				setDirty: function(dirty) {
 					var editors = that._compareView.getWidget().getEditors();
 					var checkedDirty = dirty;
-					if(editors && editors.length === 2){
-						checkedDirty = editors[0].isDirty() || editors[1].isDirty();
-					}
-					mGlobalCommands.setDirtyIndicator(checkedDirty);
-					if(that._compareView.getWidget().refreshTitle){
-						that._compareView.getWidget().refreshTitle(this._editorIndex, dirty);
+					if(!this.embedded) {
+						if(editors && editors.length === 2){
+							checkedDirty = editors[0].isDirty() || editors[1].isDirty();
+						}
+						mGlobalCommands.setDirtyIndicator(checkedDirty);
+						if(that._compareView.getWidget().refreshTitle){
+							that._compareView.getWidget().refreshTitle(this._editorIndex, dirty);
+						}
+					} else if(that._compareView.getWidget().options.titleIds && that._compareView.getWidget().options.titleIds.length > 0) {
+						var titleNode;
+						if(that._compareView.getWidget().options.maximized && that._compareView.getWidget().options.titleIds.length === 2) {
+							titleNode = lib.node(that._compareView.getWidget().options.titleIds[1]);
+							if(titleNode) {
+								titleNode.textContent = that._compareView.getWidget().options.maximized && dirty ? "*" : "";
+							}
+						}
+						
+						titleNode = lib.node(that._compareView.getWidget().options.titleIds[0]);
+						if(!titleNode) {
+							return;
+						}
+						//lib.empty(titleNode);
+						var label = titleNode.textContent;
+						if(label) {
+							if (label.charAt(label.length -1) === '*') { //$NON-NLS-0$
+								label = label.substring(0, label.length-1);
+							}
+							titleNode.textContent = checkedDirty ? label + "*" : label;
+						}
 					}
 				},
 				
@@ -251,11 +300,13 @@ exports.ResourceComparer = (function() {
 					that._progress.progress(that._fileClient.read(fileURI, true), i18nUtil.formatMessage(messages["readingFileMetadata"], fileURI)).then( //$NON-NLS-0$
 						function(metadata) {
 							this._fileMetadata = metadata;
-							var toolbar = lib.node(this._actionBarId); //$NON-NLS-0$
-							if (toolbar) {	
-								that._commandService.destroy(toolbar);
-								var editorIndex = this._editorIndex;
-								that._commandService.renderCommands(toolbar.id, toolbar, that._compareView.getWidget().getEditors()[editorIndex], that._compareView.getWidget().getEditors()[editorIndex], "tool"); //$NON-NLS-0$
+							if(!this.embedded) {
+								var toolbar = lib.node(this._actionBarId); //$NON-NLS-0$
+								if (toolbar) {	
+									that._commandService.destroy(toolbar);
+									var editorIndex = this._editorIndex;
+									that._commandService.renderCommands(toolbar.id, toolbar, that._compareView.getWidget().getEditors()[editorIndex], that._compareView.getWidget().getEditors()[editorIndex], "tool"); //$NON-NLS-0$
+								}
 							}
 							if(metadata){
 								this.setTitle(metadata.Location, metadata);
@@ -270,6 +321,9 @@ exports.ResourceComparer = (function() {
 				},
 				
 				setTitle : function(title, /*optional*/ metadata) {
+					if(this.embedded) {
+						return;
+					}
 					var name;
 					if (metadata) {
 						name = metadata.Name;
@@ -284,12 +338,27 @@ exports.ResourceComparer = (function() {
 					} 
 				},
 				
-				afterSave: function(){
+				postSave: function(closing){
 					var editors = that._compareView.getWidget().getEditors();
 					var newContents = editors[this._editorIndex].getTextView().getText();
 					var fileObj = that._getFileOptions(this._editorIndex);
 					fileObj.Content = newContents;
-					that._compareView.getWidget().refresh();
+					if(!closing) {
+						var options = that._compareView.getWidget().options;
+						if(options.diffProvider && options.diffContent) {
+							var ignoreWS = options.ignoreWhitespace ? "true" : "false";
+							options.diffProvider._diffProvider.getDiffContent(options.resource, {ignoreWS: ignoreWS}).then(function(jsonData) {
+								if (options.hasConflicts) {
+									options.diffContent = jsonData.split("diff --git")[1]; //$NON-NLS-0$
+								} else {
+									options.diffContent = jsonData;
+								}
+								that._compareView.getWidget().refresh(true, true, this.embedded ? null : this._editorIndex);
+							}.bind(this), function(){});
+						} else {
+							that._compareView.getWidget().refresh(true, true, this.embedded ? null : this._editorIndex);
+						}
+					}
 				}
 			});
 		},
@@ -335,7 +404,7 @@ exports.ResourceComparer = (function() {
 			if(cmdProvider && cmdProvider.getOptions().commandSpanId) {
 				var commandSpanId = cmdProvider.getOptions().commandSpanId;
 				var generateLinkCommand = new mCommands.Command({
-					tooltip : messages["Generate link of the current diff"],
+					tooltip : messages["GenerateCurDiffLink"],
 					name: messages["Generate Link"],
 					//imageClass : "core-sprite-link", //$NON-NLS-0$
 					id: "orion.compare.generateLink", //$NON-NLS-0$
@@ -390,6 +459,39 @@ exports.ResourceComparer = (function() {
 		        }.bind(this)
 			);
 	    },
+	    _inputChanged: function() {
+	    	var that =this;
+			this._inputManagers.forEach(function(inputManager) {
+				if(inputManager.manager){
+					var fileOptions = that._getFileOptions(inputManager.manager._editorIndex);
+					var editor = inputManager.manager.getEditor();
+					inputManager.manager.filePath = fileOptions.URL;
+					inputManager.manager.setInput(fileOptions.URL , editor);
+				}
+			});
+	    },
+	    save: function(doSave) {
+			if(doSave) {
+			   	var promises = [];
+				this._inputManagers.forEach(function(inputManager) {
+					if(inputManager.manager){
+						var editor = inputManager.manager.getEditor();
+						if(editor.isDirty()) {
+							promises.push(inputManager.manager.save(true));
+						}
+					}
+				});
+				return Deferred.all(promises);
+			} else {
+				return new Deferred().resolve();
+			}
+	    },
+		isDirty: function(){
+			return this._compareView.isDirty();
+		},
+		destroy: function(){
+			this._compareView.destroy();
+		},
 		start: function(){
 			if(this.options.resource){
 				if(!this.options.diffProvider){
@@ -397,7 +499,13 @@ exports.ResourceComparer = (function() {
 					return;
 				}
 				var that = this;
-				return that.options.diffProvider.resolveDiff(that.options.resource, that.options.compareTo, that.options.hasConflicts).then( function(diffParam){
+				return that.options.diffProvider.resolveDiff(that.options.resource, that.options.compareTo, that.options.hasConflicts, that.options.ignoreWhitespace).then( function(diffParam){
+					if(diffParam.oldFile) {
+						diffParam.oldFile.readonly = that._compareView.getWidget().options.oldFile.readonly;
+					}
+					if(diffParam.newFile) {
+						diffParam.newFile.readonly = that._compareView.getWidget().options.newFile.readonly;
+					}
 					that._compareView.getWidget().setOptions(diffParam);
 					var isImage = mContentTypes.isImage(diffParam.newFile.Type);
 					var viewOptions = that._compareView.getWidget().options;
@@ -413,17 +521,10 @@ exports.ResourceComparer = (function() {
 							return new Deferred().resolve(height + 5);
 						});
 					} else {
-						var filesToLoad = ( viewOptions.diffContent ? [viewOptions.oldFile/*, viewOptions.newFile*/] : [viewOptions.oldFile, viewOptions.newFile]); 
+						var filesToLoad = ( viewOptions.diffContent ? [viewOptions.oldFile, viewOptions.newFile] : [viewOptions.oldFile, viewOptions.newFile]); 
 						return that._getFilesContents(filesToLoad).then( function(){
 							var viewHeight = that._compareView.getWidget().refresh(true);
-							this._inputManagers.forEach(function(inputManager) {
-								if(inputManager.manager){
-									var fileOptions = that._getFileOptions(inputManager.manager._editorIndex);
-									var editor = that._compareView.getWidget().getEditors()[inputManager.manager._editorIndex];
-									inputManager.manager.filePath = fileOptions.URL;
-									inputManager.manager.setInput(fileOptions.URL , editor);
-								}
-							});
+							that._inputChanged();							
 							return new Deferred().resolve(viewHeight);
 						}.bind(that));
 					}
